requirements.txt
PyGithub==1.58.1
kubernetes==26.1.0
pyyaml==6.0
